Consolidate messaging types
