+++ dreamos/core/__init__.py
@@ -12,7 +12,7 @@ from . import messaging
 from . import persistent_queue
 from . import system_init
 from .cell_phone import CellPhone, send_message
-from .messaging.types import Message, MessageMode
+from .messaging.common import Message, MessageMode, MessagePriority
 from .persistent_queue import PersistentQueue
 
 __all__ = [
@@ -26,6 +26,7 @@ __all__ = [
     'CellPhone',
     'Message',
     'MessageMode',
+    'MessagePriority',
     'PersistentQueue',
     'send_message'
 ]
