+++ dreamos/core/messaging/__init__.py
@@ -2,13 +2,14 @@
 Messaging module for Dream.OS.
 """
 
-from .types import Message, MessageMode
+from .common import Message, MessageMode, MessagePriority
 from dreamos.core.cell_phone import CellPhone
 from dreamos.core.message_processor import MessageProcessor
 
 __all__ = [
     'Message',
     'MessageMode',
+    'MessagePriority',
     'CellPhone',
     'MessageProcessor'
 ]
